Adjust MultipleFragmentsViewModel.kt and create test page

+++ features/tokofood/src/main/java/com/tokopedia/tokofood/common/minicartwidget/view/MiniCartUiModel.kt
@@ -8,7 +8,7 @@ import kotlinx.parcelize.Parcelize
 data class MiniCartUiModel(
         var cartData: MutableMap<String, CartProduct> = mutableMapOf(),
         var shopName: String = "",
-        var totalPrice: Long = 0L,
+        var totalPrice: Double = 0.0,
         var totalProduct: Int = 0,
         var totalProductQuantity: Int = 0
 ): Parcelable
